+++ internal/resolvers/default/load.go
@@ -26,7 +26,7 @@ func (dr *R) load(urlString string, r *http.Request) (interface{}, time.Duration
 
 	for _, m := range dr.customResolvers {
 		if m.Check(requestUrl) {
-			data, err := m.Run(requestUrl)
+			data, err := m.Run(requestUrl, r)
 
 			if errors.Is(err, resolver.ErrDontHandle) {
 				break
@@ -57,7 +57,7 @@ func (dr *R) load(urlString string, r *http.Request) (interface{}, time.Duration
 	if requestUrl.String() != resp.Request.URL.String() {
 		for _, m := range dr.customResolvers {
 			if m.Check(resp.Request.URL) {
-				data, err := m.Run(resp.Request.URL)
+				data, err := m.Run(resp.Request.URL, r)
 
 				if errors.Is(err, resolver.ErrDontHandle) {
 					break