+++ internal/resolvers/imgur/run.go
@@ -3,11 +3,12 @@ package imgur
 import (
 	"encoding/json"
 	"errors"
+	"net/http"
 	"net/url"
 )
 
-func run(url *url.URL) ([]byte, error) {
-	imgurResponse, ok := imgurCache.Get(url.String(), nil).(response)
+func run(url *url.URL, r *http.Request) ([]byte, error) {
+	imgurResponse, ok := imgurCache.Get(url.String(), r).(response)
 	if !ok {
 		return nil, errors.New("imgur cache load function is broken")
 	}