+++ internal/resolvers/supinic/run.go
@@ -2,10 +2,11 @@ package supinic
 
 import (
 	"encoding/json"
+	"net/http"
 	"net/url"
 )
 
-func run(url *url.URL) ([]byte, error) {
+func run(url *url.URL, r *http.Request) ([]byte, error) {
 	matches := trackPathRegex.FindStringSubmatch(url.Path)
 	if len(matches) != 2 {
 		return nil, errInvalidTrackPath
@@ -13,6 +14,6 @@ func run(url *url.URL) ([]byte, error) {
 
 	trackID := matches[1]
 
-	apiResponse := trackListCache.Get(trackID, nil)
+	apiResponse := trackListCache.Get(trackID, r)
 	return json.Marshal(apiResponse)
 }