+++ internal/resolvers/wikipedia/run.go
@@ -3,11 +3,12 @@ package wikipedia
 import (
 	"encoding/json"
 	"errors"
+	"net/http"
 	"net/url"
 )
 
-func run(url *url.URL) ([]byte, error) {
-	wikiResponse, ok := wikipediaCache.Get(url.String(), nil).(response)
+func run(url *url.URL, r *http.Request) ([]byte, error) {
+	wikiResponse, ok := wikipediaCache.Get(url.String(), r).(response)
 	if !ok {
 		return nil, errors.New("wikipedia cache load function is broken")
 	}